requirements.txt
# Core dependencies
google-generativeai>=0.8.0
python-dotenv==1.0.1
pillow==10.4.0
pillow-heif==0.15.0

# Streamlit app only
streamlit==1.39.0
